Black_formatter added in Github Actions.
The workflow is triggered on push to the main and master branches. The workflow consists of one job that starts by checking out the repository's code and setting up a specific version of Python. Then it runs the command for installing black, and finally runs the command "black ." which formats all the python files in the current directory and subdirectories.