Create main.yml
This workflow will create a conda environment using the environment.yml file, it will install black in the environment, then it will run the black formatter on all the python files in your repository and check if any files need to be formatted. It will also run pytest and flake8 on the code in the repository.